+++ infrastructure-kotlin/src/main/kotlin/kotlin/internal/KotlinAccessors.kt
@@ -1,6 +1,7 @@
 package com.redmadrobot.build.kotlin.internal
 
 import com.redmadrobot.build.InfrastructurePlugin
+import com.redmadrobot.build.internal.InternalGradleInfrastructureApi
 import org.gradle.api.Project
 import org.gradle.api.plugins.JavaPluginExtension
 import org.gradle.kotlin.dsl.getByName
@@ -10,6 +11,7 @@ internal fun Project.java(configure: JavaPluginExtension.() -> Unit) {
     extensions.configure("java", configure)
 }
 
+@InternalGradleInfrastructureApi
 public val InfrastructurePlugin.kotlin: KotlinProjectExtension get() = project.kotlin
 
 internal val Project.kotlin: KotlinProjectExtension get() = extensions.getByName<KotlinProjectExtension>("kotlin")