+++ infrastructure-kotlin/src/main/kotlin/kotlin/internal/Test.kt
@@ -1,9 +1,11 @@
 package com.redmadrobot.build.kotlin.internal
 
+import com.redmadrobot.build.internal.InternalGradleInfrastructureApi
 import com.redmadrobot.build.kotlin.TestOptions
 import com.redmadrobot.build.kotlin.TestOptionsImpl
 import org.gradle.api.tasks.testing.Test
 
+@InternalGradleInfrastructureApi
 public fun Test.setTestOptions(testOptions: TestOptions) {
     val configure = (testOptions as TestOptionsImpl).configuration.get()
     if (testOptions.useJunitPlatform.get()) {