+++ infrastructure-publish/src/main/kotlin/publish/PublishConfigPlugin.kt
@@ -1,6 +1,7 @@
 package com.redmadrobot.build.publish
 
 import com.redmadrobot.build.InfrastructurePlugin
+import com.redmadrobot.build.internal.InternalGradleInfrastructureApi
 import org.gradle.api.Project
 
 /**
@@ -14,6 +15,7 @@ public open class PublishConfigPlugin : InfrastructurePlugin() {
     internal lateinit var publishingOptions: PublishingOptionsImpl
         private set
 
+    @InternalGradleInfrastructureApi
     override fun Project.configure() {
         publishingOptions = createExtension("publishing")
     }